Format and Move the Utility Module
Moves the Utility modules out of the Node.

Cuts old functionality which is currently not used by the codebase.
