+++ documentation/analysis/fema-analysis-pinger.livemd
@@ -77,7 +77,7 @@ Kino.Process.render_seq_trace(
     # we should use the router, but pinger is special
     Pinger.set_timer(node.pinger, 20)
   end,
-  message_label: &Anoma.Node.Utility.message_label/1
+  message_label: &Anoma.Utility.message_label/1
 )
 ```
 
@@ -114,7 +114,7 @@ Kino.Process.render_seq_trace(
     send(Process.whereis(node.pinger.server), :execute)
     :timer.sleep(1)
   end,
-  message_label: &Anoma.Node.Utility.message_label/1
+  message_label: &Anoma.Utility.message_label/1
 )
 ```
