+++ documentation/visualization/actors.livemd
@@ -118,7 +118,7 @@ Kino.Process.render_seq_trace(
   fn ->
     Mempool.execute(node.mempool)
   end,
-  message_label: &Anoma.Node.Utility.message_label/1
+  message_label: &Anoma.Utility.message_label/1
 )
 ```
 
@@ -186,7 +186,7 @@ We can also see what processes startup when we start an execution
 Kino.Process.render_seq_trace(
   [Process.whereis(node.mempool.server)],
   fn -> Mempool.tx(node.mempool, {:kv, increment_counter_val(555)}).pid() end,
-  message_label: &Anoma.Node.Utility.message_label/1
+  message_label: &Anoma.Utility.message_label/1
 )
 ```
 
@@ -200,7 +200,7 @@ Kino.Process.render_seq_trace(
 Kino.Process.render_seq_trace(
   :all,
   fn -> Anoma.Node.Logger.add(node.logger, :info, "help") end,
-  message_label: &Anoma.Node.Utility.message_label/1
+  message_label: &Anoma.Utility.message_label/1
 )
 ```
